Refactor preset data and functions to preset module
