+++ pretenders/boss/views/preset.py
@@ -3,11 +3,24 @@ import re
 import bottle
 from bottle import delete, post, HTTPResponse
 
+try:
+    from collections import OrderedDict
+except ImportError:
+    #2.6 compatibility
+    from pretenders.compat.ordered_dict import OrderedDict
+
 from pretenders.base import get_logger
-from pretenders.boss import data
 from pretenders.http import Preset
 
 LOGGER = get_logger('pretenders.boss.views.preset')
+PRESETS = OrderedDict()
+
+
+def preset_count():
+    """
+    Check whether there are any presets.
+    """
+    return len(PRESETS)
 
 
 def select_preset(value):
@@ -23,15 +36,21 @@ def select_preset(value):
 
     Return 404 if no preset found that matches.
     """
-    for key, preset_list in data.PRESETS.items():
+    for key, preset_list in PRESETS.items():
 
         preset = preset_list[0]
         preset_matches = re.match(preset.rule, value)
         if preset_matches:
-            data.pop_preset(preset_list, key)
+            pop_preset(preset_list, key)
             return preset
 
     raise HTTPResponse(b"No matching preset response", status=404)
+
+
+def pop_preset(preset_list, key):
+    del preset_list[0]
+    if not preset_list:
+        del PRESETS[key]
 
 
 @post('/preset')
@@ -41,9 +60,9 @@ def add_preset():
     """
     preset = Preset(json_data=bottle.request.body.read())
     rule = preset.rule
-    if rule not in data.PRESETS:
-        data.PRESETS[rule] = []
-    url_presets = data.PRESETS[rule]
+    if rule not in PRESETS:
+        PRESETS[rule] = []
+    url_presets = PRESETS[rule]
     url_presets.append(preset)
 
 
@@ -52,4 +71,4 @@ def clear_presets():
     """
     Delete all recorded presets
     """
-    data.PRESETS.clear()
+    PRESETS.clear()
